Rebase history to remove LiABS
Due to a change in development roadmap, some files and their history included functions that we're not ready to share yet. This led to removing the file from history to protect that information. Unfortunately, ABS-related commits before this one might look a bit strange as history has been rewritten and if they edited the abs.cpp file or the ABSClass.R file this is no longer shown. The next commits reintroduce these files